Gh019 read modulemaps with parser (#29)
Update spm_repository to use the Starlark module map parser when a custom module map file is provided.
Added clang_module_headers to spm_package rule so that header specified in a custom module map file can be specified.
The public header files for all clang targets are copied to the output.
A custom module.modulemap file is written for all clang targets pointing to one of the public headers as an umbrella header (i.e. include the specified header and all other headers in the directory and sub-directories).