Fixed inability to clone git repositories in SPM build (#30)
Perform the SPM package resolve/fetch in spm_repository. Then, in spm_package, we copy the fetched files over to the build path.
Changed the vapor example to package_with_deps using AsyncKit.
Example fails to build because deps still point to sandboxed headers. Need to apply the modulemap magic that we apply to top-level packages.