Transitioned to Swift toolchain and added an arch flag to SPM build script (#44)
Removed the Bazel toolchain implementation and replaced it with consumption of the SwiftToolchainInfo from rules_swift. This ensures that the SPM build will build with the correct platform information.
Updated the SPM build script to accept flags. The previous implementation relied solely on positional arguments which made the script brittle and harder to grok.
Added an --arch flag to the SPM build script so that the proper architecture is built.