Get `rules_spm` working on Linux (#52)
Reworked spm_clang_module and spm_swift_module to not use objc_library. This allows the code to work on macos and linux.
Update the SPM build script to support the passing of the Swift executable and an optional SDK name.
Added spm_archive rule. It creates an archive (.a) file from one or more object files (.o).
Added spm_autoconfiguration to determine which toolchain to use for the build. See Update toolchain evaluation to use Bazel toolchain evaluation #50 for more details.
Updated the Vapor example to not use Apple-specific rules.
Updated the README.md to include Linux prerequisite