Generate documentation. (#71)
Added and update bzl_library declarations for all of the items that are being documented.
Added scripts to generate the documentation. (bazel run //doc:update).
Added diff_test declarations to ensure that the latest code has been generated.
Cleaned up a bunch of the existing doc strings to ensure that it works with stardoc.
Added links to the new doc.