Parse paths as set rather than list.
Occasionally a directory will be included multiple times.
This can happen when the includes is underneath the root path for the sources.
e.g. https://github.com/TimOliver/TOCropViewController/tree/main/Objective-C/TOCropViewController
In many cases this may be acceptable. But if a module.modulemap file
is included twice, we don't want it to fail.